moving feeddiagnostics to implementation
